@@ -1,246 +1,258 @@
-module.exports = {
-	"rules": {
-		/* eslint-enable sort-keys */
+export default {
+	// https://github.com/eslint/eslint @ v8.19.0
+	rules: {
 		"accessor-pairs": "off",
-		"array-bracket-newline": "off",
-		"array-bracket-spacing": "off",
-		"array-callback-return": "off",
-		"array-element-newline": "off",
-		"arrow-body-style": "off",
+		"array-bracket-newline": ["error", "consistent"],
+		"array-bracket-spacing": ["error", "never"],
+		"array-callback-return": ["error", {"checkForEach": true}],
+		"array-element-newline": ["error", "consistent"],
+		"arrow-body-style": ["error", "as-needed"],
 		"arrow-parens": ["error", "always"],
 		"arrow-spacing": ["error", {
 			"before": true,
 			"after": true
 		}],
-		"block-scoped-var": "off",
-		"block-spacing": "off",
-		"brace-style": "error",
-		"callback-return": "off",
-		"camelcase": ["error", {
-			"properties": "never"
-		}],
+		"block-scoped-var": "error",
+		"block-spacing": ["error", "never"],
+		"brace-style": ["error", "1tbs"],
+		"camelcase": "error",
 		"capitalized-comments": "off",
 		"class-methods-use-this": "off",
-		"comma-dangle": ["warn", "never"],
+		"comma-dangle": ["error", "never"],
 		"comma-spacing": ["error", {
 			"before": false,
 			"after": true
-		}],
+		}], 
 		"comma-style": ["error", "last"],
 		"complexity": "off",
-		"computed-property-spacing": ["error", "never"],
+		"computed-property-spacing": ["error", "never", {"enforceForClassMembers": true}],
 		"consistent-return": "off",
-		"consistent-this": "off",
+		"consistent-this": "error",
 		"constructor-super": "error",
 		"curly": "error",
+		"default-case-last": "error",
 		"default-case": "off",
-		"dot-location": "off",
-		"dot-notation": "off",
+		"default-param-last": "error",
+		"dot-location": ["error", "property"],
+		"dot-notation": "error",
 		"eol-last": ["error", "always"],
 		"eqeqeq": "error",
 		"for-direction": "off",
 		"func-call-spacing": ["error", "never"],
-		"func-name-matching": "off",
+		"func-name-matching": "error",
 		"func-names": "off",
-		"func-style": "off",
-		"generator-star-spacing": "off",
-		"global-require": "off",
+		"func-style": ["error", "declaration", {"allowArrowFunctions": true}],
+		"function-call-argument-newline": ["error", "consistent"],
+		"function-paren-newline": ["error", "consistent"],
+		"generator-star-spacing": ["error", {"before": true, "after": false}],
+		"getter-return": "error",
+		"grouped-accessor-pairs": ["error", "getBeforeSet"],
 		"guard-for-in": "error",
-		"handle-callback-err": "error",
-		"id-blacklist": "off",
+		"id-denylist": "off",
 		"id-length": "off",
 		"id-match": "off",
-		"indent": ["warn", "tab"],
-		"indent-legacy": "off",
+		"implicit-arrow-linebreak": ["error", "beside"],
+		"indent": ["error", "tab"],
 		"init-declarations": "off",
-		"jsx-quotes": "off",
-		"key-spacing":  ["error", {
+		"jsx-quotes": ["error", "prefer-double"],
+		"key-spacing": ["error", {
 			"afterColon": true,
 			"beforeColon": false,
 			"mode": "strict"
 		}],
 		"keyword-spacing": ["error", {
-			"before": true
+			"before": true,
+			"after": true
 		}],
-		"line-comment-position": "off",
+		"line-comment-position": ["error", {"position": "above"}],
 		"linebreak-style": ["error", "unix"],
 		"lines-around-comment": "off",
-		"lines-around-directive": "off",
-		"max-depth": "off",
-		"max-len": "off",
-		"max-lines": "off",
-		"max-nested-callbacks": "off",
+		"lines-between-class-members": "off",
+		"max-classes-per-file": "off",
+		"max-depth": ["error", 4],
+		"max-len": ["error", {"code": 80, "tabWidth": 4}],
+		"max-lines-per-function": ["error", {"max": 50}],
+		"max-lines": ["error", 300],
+		"max-nested-callbacks": ["error", 3],
 		"max-params": "off",
-		"max-statements": "off",
-		"max-statements-per-line": "off",
-		"multiline-ternary": "off",
-		"new-cap": "off",
+		"max-statements-per-line": ["error", {"max": 1}],
+		"max-statements": ["error", 10],
+		"multiline-comment-style": ["error", "separate-lines"],
+		"multiline-ternary": ["error", "always-multiline"],
+		"new-cap": ["error", {"newIsCap": true, "capIsNew": true, "properties": true}],
 		"new-parens": "error",
-		"newline-after-var": "off",
-		"newline-before-return": "off",
-		"newline-per-chained-call": "off",
-		"no-alert": "off",
+		"newline-per-chained-call": "error",
+		"no-alert": "error",
 		"no-array-constructor": "error",
-		"no-await-in-loop": "off",
-		"no-bitwise": "off",
-		"no-buffer-constructor": "off",
-		"no-caller": "off",
+		"no-async-promise-executor": "error",
+		"no-await-in-loop": "error",
+		"no-bitwise": "error",
+		"no-caller": "error",
 		"no-case-declarations": "error",
-		"no-catch-shadow": "off",
 		"no-class-assign": "error",
 		"no-compare-neg-zero": "error",
 		"no-cond-assign": "error",
 		"no-confusing-arrow": ["error", {
-			"allowParens": false
+			"allowParens": true
 		}],
-		"no-console": "off",
+		"no-console": ["error", {allow: ["warn", "error"]}],
 		"no-const-assign": "error",
-		"no-constant-condition": "warn",
+		"no-constant-binary-expression": "error",
+		"no-constant-condition": ["error", {"checkLoops": false}],
+		"no-constructor-return": "error",
 		"no-continue": "off",
 		"no-control-regex": "error",
-		"no-debugger": "warn",
+		"no-debugger": "error",
 		"no-delete-var": "error",
-		"no-div-regex": "off",
+		"no-div-regex": "error",
 		"no-dupe-args": "error",
 		"no-dupe-class-members": "error",
+		"no-dupe-else-if": "error",
 		"no-dupe-keys": "error",
 		"no-duplicate-case": "error",
 		"no-duplicate-imports": "error",
-		"no-else-return": "error",
-		"no-empty": "error",
+		"no-else-return": ["error", {"allowElseIf": true}],
 		"no-empty-character-class": "error",
-		"no-empty-function": "off",
+		"no-empty-function": "error",
 		"no-empty-pattern": "error",
-		"no-eq-null": "off",
-		"no-eval": "off",
+		"no-empty": "error",
+		"no-eq-null": "error",
+		"no-eval": "error",
 		"no-ex-assign": "error",
-		"no-extend-native": "off",
-		"no-extra-bind": "off",
-		"no-extra-boolean-cast": "error",
-		"no-extra-label": "off",
-		"no-extra-parens": "off",
+		"no-extend-native": "error",
+		"no-extra-bind": "error",
+		"no-extra-boolean-cast": ["error", {"enforceForLogicalOperands": true}],
+		"no-extra-label": "error",
+		"no-extra-parens": ["error", "all", {"ignoreJSX": "multi-line"}],
 		"no-extra-semi": "error",
 		"no-fallthrough": "error",
-		"no-floating-decimal": "off",
+		"no-floating-decimal": "error",
 		"no-func-assign": "error",
 		"no-global-assign": "error",
-		"no-implicit-coercion": "off",
-		"no-implicit-globals": "error",
-		"no-implied-eval": "off",
-		"no-inline-comments": "off",
+		"no-implicit-coercion": "error",
+		"no-implicit-globals": ["error", {"lexicalBindings": true}],
+		"no-implied-eval": "error",
+		"no-import-assign": "error",
+		"no-inline-comments": "error",
 		"no-inner-declarations": "error",
 		"no-invalid-regexp": "error",
-		"no-invalid-this": "off",
-		"no-irregular-whitespace": "error",
-		"no-iterator": "off",
-		"no-label-var": "off",
-		"no-labels": "off",
-		"no-lone-blocks": "off",
-		"no-lonely-if": "warn",
-		"no-loop-func": "off",
+		"no-invalid-this": "error",
+		"no-irregular-whitespace": ["error", {"skipStrings": false}],
+		"no-iterator": "error",
+		"no-label-var": "error",
+		"no-labels": "error",
+		"no-lone-blocks": "error",
+		"no-lonely-if": "error",
+		"no-loop-func": "error",
+		"no-loss-of-precision": "error",
 		"no-magic-numbers": "off",
+		"no-misleading-character-class": "error",
 		"no-mixed-operators": "error",
-		"no-mixed-requires": "off",
-		"no-mixed-spaces-and-tabs": ["warn", "smart-tabs"],
+		"no-mixed-spaces-and-tabs": "error",
 		"no-multi-assign": "error",
 		"no-multi-spaces": "error",
-		"no-multi-str": "off",
-		"no-multiple-empty-lines": "off",
-		"no-native-reassign": "off",
+		"no-multi-str": "error",
+		"no-multiple-empty-lines": ["error", {"max": 3, "maxBOF": 0, "maxEOF": 1}],
 		"no-negated-condition": "off",
-		"no-negated-in-lhs": "off",
 		"no-nested-ternary": "off",
-		"no-new": "off",
-		"no-new-func": "off",
+		"no-new-func": "error",
 		"no-new-object": "error",
-		"no-new-require": "off",
 		"no-new-symbol": "error",
 		"no-new-wrappers": "error",
+		"no-new": "error",
+		"no-nonoctal-decimal-escape": "error",
 		"no-obj-calls": "error",
+		"no-octal-escape": "error",
 		"no-octal": "error",
-		"no-octal-escape": "off",
-		"no-param-reassign": "off",
-		"no-path-concat": "error",
+		"no-param-reassign": ["error", {"props": false}],
 		"no-plusplus": "off",
-		"no-process-env": "off",
-		"no-process-exit": "off",
-		"no-proto": "off",
+		"no-promise-executor-return": "error",
+		"no-proto": "error",
 		"no-prototype-builtins": "off",
 		"no-redeclare": "error",
 		"no-regex-spaces": "error",
+		"no-restricted-exports": "off",
 		"no-restricted-globals": "off",
 		"no-restricted-imports": "off",
-		"no-restricted-modules": "off",
 		"no-restricted-properties": "off",
 		"no-restricted-syntax": "off",
-		"no-return-assign": "error",
+		"no-return-assign": ["error", "always"],
 		"no-return-await": "off",
-		"no-script-url": "off",
+		"no-script-url": "error",
 		"no-self-assign": "error",
-		"no-self-compare": "off",
-		"no-sequences": "off",
-		"no-shadow": "off",
-		"no-shadow-restricted-names": "off",
-		"no-spaced-func": "off",
+		"no-self-compare": "error",
+		"no-sequences": ["error", {"allowInParentheses": false}],
+		"no-setter-return": "error",
+		"no-shadow-restricted-names": "error",
+		"no-shadow": ["error", {"builtinGlobals": true, "hoist": "all"}],
 		"no-sparse-arrays": "error",
-		"no-sync": "off",
-		"no-tabs": "off",
-		"no-template-curly-in-string": "off",
+		"no-tabs": ["error", {"allowIndentationTabs": true}],
+		"no-template-curly-in-string": "error",
 		"no-ternary": "off",
 		"no-this-before-super": "error",
 		"no-throw-literal": "error",
 		"no-trailing-spaces": "error",
+		"no-undef-init": "error",
 		"no-undef": "error",
-		"no-undef-init": "off",
-		"no-undefined": "off",
+		"no-undefined": "error",
 		"no-underscore-dangle": "off",
 		"no-unexpected-multiline": "error",
-		"no-unmodified-loop-condition": "off",
-		"no-unneeded-ternary": "error",
-		"no-unreachable": "warn",
+		"no-unmodified-loop-condition": "error",
+		"no-unneeded-ternary": ["error", {"defaultAssignment": false}],
+		"no-unreachable-loop": "error",
+		"no-unreachable": "error",
 		"no-unsafe-finally": "error",
 		"no-unsafe-negation": "error",
-		"no-unused-expressions": "error",
+		"no-unsafe-optional-chaining": ["error", {"disallowArithmeticOperators": true}],
+		"no-unused-expressions": ["error", {"enforceForJSX": true}],
 		"no-unused-labels": "error",
-		"no-unused-vars": "warn",
-		"no-use-before-define": "off",
-		"no-useless-call": "off",
-		"no-useless-computed-key": "error",
-		"no-useless-concat": "off",
+		"no-unused-private-class-members": "error",
+		"no-unused-vars": "error",
+		"no-use-before-define": "error",
+		"no-useless-backreference": "error",
+		"no-useless-call": "error",
+		"no-useless-catch": "error",
+		"no-useless-computed-key": ["error", {"enforceForClassMembers": true}],
+		"no-useless-concat": "error",
 		"no-useless-constructor": "error",
 		"no-useless-escape": "error",
 		"no-useless-rename": "error",
-		"no-useless-return": "off",
+		"no-useless-return": "error",
 		"no-var": "error",
-		"no-void": "off",
+		"no-void": "error",
 		"no-warning-comments": "off",
 		"no-whitespace-before-property": "error",
-		"no-with": "off",
-		"nonblock-statement-body-position": "off",
-		"object-curly-newline": "off",
+		"no-with": "error",
+		"nonblock-statement-body-position": ["error", "below"],
+		"object-curly-newline": ["error", {"multiline": true}],
 		"object-curly-spacing": ["error", "never"],
 		"object-property-newline": "off",
-		"object-shorthand": "off",
+		"object-shorthand": ["error", "always"],
+		"one-var-declaration-per-line": ["error", "initializations"],
 		"one-var": "off",
-		"one-var-declaration-per-line": "off",
-		"operator-assignment": "error",
+		"operator-assignment": ["error", "always"],
 		"operator-linebreak": ["error", "after"],
-		"padded-blocks": ["warn", "never"],
+		"padded-blocks": ["error", "never"],
 		"padding-line-between-statements": "off",
-		"prefer-arrow-callback": "off",
+		"prefer-arrow-callback": "error",
 		"prefer-const": "error",
 		"prefer-destructuring": "off",
+		"prefer-exponentiation-operator": "error",
+		"prefer-named-capture-group": "error",
 		"prefer-numeric-literals": "off",
+		"prefer-object-has-own": "error",
+		"prefer-object-spread": "error",
 		"prefer-promise-reject-errors": "error",
-		"prefer-reflect": "off",
-		"prefer-rest-params": "off",
-		"prefer-spread": "off",
+		"prefer-regex-literals": ["error", {"disallowRedundantWrapping": true}],
+		"prefer-rest-params": "error",
+		"prefer-spread": "error",
 		"prefer-template": "error",
 		"quote-props": ["error", "consistent-as-needed"],
 		"quotes": ["error", "single"],
-		"radix": "error",
-		"require-await": "warn",
-		"require-jsdoc": "off",
+		"radix": ["error", "as-needed"],
+		"require-atomic-updates": "error",
+		"require-await": "error",
+		"require-unicode-regexp": "error",
 		"require-yield": "error",
 		"rest-spread-spacing": ["error", "never"],
 		"semi": ["error", "always"],
@@ -248,36 +260,35 @@ module.exports = {
 			"before": false,
 			"after": true
 		}],
-		"semi-style": "off",
+		"semi-style": ["error", "last"],
 		"sort-imports": "off",
-		"sort-keys": "off",
-		"sort-vars": "off",
-		"space-before-blocks": "error",
+		"sort-keys": ["error", "asc", {"natural": true}],
+		"sort-vars": "error",
+		"space-before-blocks": ["error", "always"],
 		"space-before-function-paren": ["error", {
-			"anonymous": "always",
+			"anonymous": "never",
 			"named": "never",
-			"asyncArrow": "ignore"
-		}],
+			"asyncArrow": "always"
+		}], 
 		"space-in-parens": ["error", "never"],
-		"space-infix-ops": "error",
-		"space-unary-ops":  [2, {
+		"space-infix-ops": "off",
+		"space-unary-ops": ["error", {
 			"words": true,
 			"nonwords": false
 		}],
-		"spaced-comment": "off",
+		"spaced-comment": ["error", "always"],
 		"strict": "error",
-		"switch-colon-spacing": "off",
-		"symbol-description": "off",
-		"template-curly-spacing": "error",
-		"template-tag-spacing": "off",
+		"switch-colon-spacing": ["error", {"after": true, "before": false}],
+		"symbol-description": "error",
+		"template-curly-spacing": ["error", "never"],
+		"template-tag-spacing": ["error", "never"],
 		"unicode-bom": "error",
 		"use-isnan": "error",
-		"valid-jsdoc": "off",
 		"valid-typeof": "error",
-		"vars-on-top": "off",
-		"wrap-iife": "off",
+		"vars-on-top": "error",
+		"wrap-iife": ["error", "inside"],
 		"wrap-regex": "off",
-		"yield-star-spacing": "off",
-		"yoda": "off"
+		"yield-star-spacing": ["error", "before"],
+		"yoda": ["error", "never", { "exceptRange": true }]
 	}
-}
+};
